fix: Fixed vc stats events
